Remove C dependency for generating quine.
Actually, we don't need to resort to C to generate
the quine. And it even more easier to code this in
Brainf*ck. Now only the the JIT interpret depends
on C right now, Hooray! (seems it is quite impossible
writing the interpreter totally in itself, because BF
itself isn't capable of organizing the runtime.)